tissue sheet creating and full incorporation
Inclusion of the tissue sheet.

Each row in the tissue sheet is an independently stored tissue specimen.  One individual can have several tissue specimens.   The tissue sheet contains information on each tissue:  where it is, what it is, how it is stored, and does it still exist. Incorporate a tissue sheet using the following model: Individuals - Tissues - Extractions.