Fix species IDs for Equulites leuciscus and laterofenestra
These species are hard to tell apart. I updated the individual, lot, sequence info, and species sheets to reflect what we know about individual ID. Clark et al. 2025 PNAS should be able to tell us which individuals in lot 191979 are which species, but I did not yet include this information.